+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Add experimental native backend support for Windows and Linux ([#1295](https://github.com/getsentry/sentry-unreal/pull/1295))
+- Add automatic frame time metrics ([#1293](https://github.com/getsentry/sentry-unreal/pull/1293))
 
 ### Dependencies
 
@@ -17,6 +18,9 @@
 - Bump Native SDK from v0.13.2 to v0.13.3 ([#1299](https://github.com/getsentry/sentry-unreal/pull/1299))
   - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#0133)
   - [diff](https://github.com/getsentry/sentry-native/compare/0.13.2...0.13.3)
+- Bump Cocoa SDK from v9.7.0 to v9.8.0 ([#1298](https://github.com/getsentry/sentry-unreal/pull/1298))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#980)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/9.7.0...9.8.0)
 
 ## 1.9.1